- Refactored apple/object architecture to allow for different object subclasses
