+++ Snake/Obstacle.py
@@ -1,2 +1,6 @@
 import Global as g
+import random
 
+class Obstacle:
+    def __init__(self,) -> None:
+        self.
